♻️ refactor(services): implement BaseResultsService for shared logic in InnResultsService and OicrResultsService, and update select component for improved type safety and functionality.
