+++ research-indicators/src/app/shared/components/all-modals/modals-content/create-result-modal/components/create-oicr-form/create-oicr-form.component.html
@@ -143,6 +143,7 @@
                 optionLabel="title"
                 serviceName="oicrResults"
                 placeholder="Select OICR"
+                scrollHeight="295px"
                 [hideSelected]="false"
                 textSpan="OICR selected">
                 <ng-template #item let-item>
